+++ include/xsk/gsc/common/assembly.hpp
@@ -222,8 +222,8 @@ enum class opcode : u8
 
 struct sourcepos
 {
-    u16 line;
-    u16 column;
+    u32 line;
+    u32 column;
 };
 
 struct instruction
